pyproject.toml
[project]
name = "rm-mcp"
version = "0.1.0"
description = "MCP server for accessing reMarkable tablet data"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "Thijs Verreck" }
]
keywords = ["mcp", "remarkable", "tablet", "notes"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "mcp>=1.0.0",
    "rmscene>=0.6.0",
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
    "rmc>=0.3.0",
    "requests>=2.28.0",
    "cairosvg>=2.8.2",
    "pymupdf>=1.24.0",
    "ebooklib>=0.18",
    "beautifulsoup4>=4.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=1.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
ocr = [
    "google-cloud-vision>=3.0.0",
]

[project.scripts]
rm-mcp = "rm_mcp.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["rm_mcp"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.6",
]
